* Fixed Comments, Moved Project Components to New Directory
