+++ pages/api/projects/[id].ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { prisma } from '../../../prisma/adapter';
 import { sendApiError, sendApiSuccess, type ApiEnvelope } from 'app/helpers/response';
+import { getProjectById } from 'app/repositories/projects';
 import { transformProject } from 'app/transformers/projects';
 
 type ProjectResponse = ApiEnvelope<ReturnType<typeof transformProject>>;
@@ -12,43 +12,9 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
       return sendApiError(res, 400, 'Invalid project id');
     }
     try {
-      const project = await prisma.project.findUnique({
-        where: { id },
-        include: {
-          job: {
-            take: 1,
-            orderBy: { priority: 'asc' },
-            include: {
-              job: {
-                include: {
-                  company: true,
-                  roles: {
-                    orderBy: [
-                      { priority: 'asc' },
-                      { start_date: 'asc' },
-                    ],
-                  },
-                  impacts: {
-                    orderBy: { priority: 'asc' },
-                  },
-                },
-              },
-            },
-          },
-          links: { orderBy: { priority: 'asc' } },
-          gallery: { orderBy: { priority: 'asc' } },
-          skills: {
-            include: { skill: true },
-            orderBy: { priority: 'asc' },
-          },
-          categories: {
-            include: { category: true },
-            orderBy: { priority: 'asc' },
-          },
-        },
-      });
+      const project = await getProjectById(id);
       if (!project) return sendApiError(res, 404, 'Project not found');
-      return sendApiSuccess(res, 200, transformProject(project));
+      return sendApiSuccess(res, 200, project);
     } catch (error) {
       console.error('GET /api/projects/[id] failed', error);
       return sendApiError(res, 500, 'Failed to fetch project');
