+++ pages/projects/[id].tsx
@@ -1,176 +1,10 @@
 import { GetServerSideProps } from "next";
-import Head from "next/head";
-import Link from "next/link";
-import { buildDateRange, toSecureAssetUrl } from "app/helpers/common";
 import type { ProjectDetail } from "app/models/projects";
 import { getProjectById } from "app/repositories/projects";
+import { ProjectDetailView } from "components/projects/ProjectDetailView";
 
 interface Props {
   project: ProjectDetail;
-}
-
-export function ProjectDetailView({ project }: Props) {
-  const isExperienceEntry = project.categories.some((categoryEntry) => categoryEntry.shortcode === 'experience');
-  const dateRange = buildDateRange(project.start_date, project.end_date);
-  const parentJob = project.job ?? null;
-  const parentJobHref = parentJob?.shortcode ? `/experience/${parentJob.shortcode}` : null;
-
-  return (
-    <>
-      <Head>
-        <title>{`${project.name} | JRProgramming`}</title>
-        <meta name="description" content={project.short} />
-      </Head>
-      <main className="min-h-screen px-4 py-12">
-        <section className="w-full max-w-4xl mx-auto">
-          <nav className="mb-4 text-sm text-primary-text/65">
-            <Link href="/" className="hover:text-primary-accentLight">Home</Link>
-            <span className="px-2 text-primary-text/40">/</span>
-            <Link href={isExperienceEntry ? '/experience' : '/projects'} className="hover:text-primary-accentLight">
-              {isExperienceEntry ? 'Experience' : 'Portfolio'}
-            </Link>
-            {parentJob && parentJobHref && (
-              <>
-                <span className="px-2 text-primary-text/40">/</span>
-                <Link href={parentJobHref} className="hover:text-primary-accentLight">
-                  {parentJob.company?.name || 'Role Breakdown'}
-                </Link>
-              </>
-            )}
-            <span className="px-2 text-primary-text/40">/</span>
-            <span className="text-primary-text/85">{project.name}</span>
-          </nav>
-
-          <div className="terminal-card mb-8 px-6 pb-8 pt-14 md:px-8">
-            <div className="mt-4 flex flex-wrap justify-between items-start gap-2 mb-2">
-              <h1 className="text-4xl md:text-5xl font-extrabold gradient-text animate-gradient">
-                {project.name}
-              </h1>
-              {dateRange && (
-                <span className="rounded-full border border-primary-accent/25 px-3 py-1 text-sm text-primary-text/65 whitespace-nowrap">{dateRange}</span>
-              )}
-            </div>
-            {project.role && (
-              <p className="text-lg text-primary-accentLight font-medium">{project.role}</p>
-            )}
-            {project.position && (
-              <p className="mt-1 text-sm text-primary-text/70">{project.position}</p>
-            )}
-          </div>
-
-          <div className="terminal-card project-block-emphasis p-6 mb-6">
-            <h2 className="mb-3 text-lg font-semibold text-accent">What It Does</h2>
-            <p className="text-text leading-relaxed">{project.short}</p>
-          </div>
-
-          <h2 className="mb-4 text-lg font-semibold text-accent">Project Snapshot</h2>
-          <div className="grid gap-5 md:grid-cols-2 mb-6">
-            <div className="terminal-card project-block-emphasis p-6">
-              <p className="text-xs uppercase tracking-[0.24em] text-primary-accentLight">Built For</p>
-              <p className="mt-2 text-sm leading-7 text-primary-text/80">{project.position || 'Client project'}</p>
-            </div>
-
-            <div className="terminal-card project-block-emphasis p-6">
-              <p className="text-xs uppercase tracking-[0.24em] text-primary-accentLight">Focus</p>
-              <p className="mt-2 text-sm leading-7 text-primary-text/80">{project.role || 'Custom software development'}</p>
-            </div>
-
-            <div className="terminal-card project-block-emphasis p-6">
-              <p className="text-xs uppercase tracking-[0.24em] text-primary-accentLight">Timeline</p>
-              <p className="mt-2 text-sm leading-7 text-primary-text/80">{dateRange || 'Timeline not specified'}</p>
-            </div>
-
-            <div className="terminal-card project-block-emphasis p-6">
-              <p className="text-xs uppercase tracking-[0.24em] text-primary-accentLight">Categories</p>
-              <div className="mt-2 flex flex-wrap gap-2">
-                {project.categories.length > 0 ? project.categories.map((categoryEntry) => (
-                  <Link
-                    key={categoryEntry.id}
-                    href={`/${categoryEntry.shortcode}`}
-                    className="rounded-full border border-accent/20 px-3 py-1 text-xs text-primary-text/70 transition hover:border-accent hover:text-accent"
-                  >
-                    {categoryEntry.title}
-                  </Link>
-                )) : (
-                  <span className="text-sm text-primary-text/60">Uncategorized</span>
-                )}
-              </div>
-            </div>
-          </div>
-
-          {project.extended && (
-            <div className="terminal-card project-block-emphasis p-6 mb-6">
-              <h2 className="text-lg font-semibold text-accent mb-3">What I Built</h2>
-              <div
-                className="prose prose-invert max-w-none text-text leading-relaxed"
-                dangerouslySetInnerHTML={{ __html: project.extended }}
-              />
-            </div>
-          )}
-
-          {project.links.length > 0 && (
-            <div className="terminal-card project-block-emphasis p-6 mb-6">
-              <h2 className="text-lg font-semibold text-accent mb-3">Proof & Links</h2>
-              <div className="flex flex-wrap gap-3">
-                {project.links.map((link) => (
-                  <a
-                    key={link.id}
-                    href={toSecureAssetUrl(link.url)}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="btn-cta-outline inline-block px-5 py-2 text-sm font-medium"
-                  >
-                    {link.website}
-                  </a>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {project.skills.length > 0 && (
-            <div className="terminal-card project-block-emphasis p-6 mb-6">
-              <h2 className="text-lg font-semibold text-accent mb-3">Skills & Technologies</h2>
-              <div className="flex flex-wrap gap-2">
-                {project.skills.map((s) => (
-                  <span
-                    key={s.id}
-                    className="px-3 py-1 rounded-full text-sm glass border border-accent/20 text-primary-accentLight"
-                    title={s.desc}
-                  >
-                    {s.name}
-                  </span>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {project.gallery.length > 0 && (
-            <div className="terminal-card project-block-emphasis p-6 mb-6">
-              <h2 className="text-lg font-semibold text-accent mb-3">Gallery</h2>
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-                {project.gallery.map((item) => (
-                  <figure key={item.id} className="rounded-lg overflow-hidden border border-accent/20">
-                    <img
-                      src={toSecureAssetUrl(item.image)}
-                      alt={item.title}
-                      className="w-full object-cover"
-                      loading="lazy"
-                    />
-                    {item.title && (
-                      <figcaption className="px-3 py-2 text-xs text-muted bg-black/30">
-                        {item.title}
-                      </figcaption>
-                    )}
-                  </figure>
-                ))}
-              </div>
-            </div>
-          )}
-
-        </section>
-      </main>
-    </>
-  );
 }
 
 export default function ProjectPage({ project }: Props) {
